Client's terminal preferences saving updates
In client mangrc saving functions was a direct call to main-x11 function
to collect position, dimensions ant font for window to save to
preferences file. This could be troublesome sometimes in the future and
is not needed now, because all the info is already stored in the
term_data.

Now the saving functions were reverted back to save only data in
term_prefs structure and any client will have to modify them to change
the preferences written to a mangrc file.

And because the preferences saving fuction is now called after terminal
nuking, the OS specific client doesn't have to chage
the term_prefs struct after every window change, but just after
nuke_hook is called (like it is now in x11 client).

Also snprintf is used for font saving to mangrc to not produce a
Wformat-overflow warning.