Move x11 terminal window initialization to a function with index parameter
In init_x11(), every terminal window was created with it's own piece of
code. Each piece contained a lot of repetitions with some differences.
The differences were agregated to x11_terms_... arrays and the 8 blocks
of code were simplified to one for cycle, calling the newly introduced
x11_term_init() function.

This was done to be able to use the x11_term_init() function later for
live toggling windows feature.